+++ api-gate/src/main/java/com/llzw/apigate/web/controller/PaymentController.java
@@ -1,11 +1,11 @@
 package com.llzw.apigate.web.controller;
 
+import com.llzw.apigate.message.RestResponseFactory;
 import com.llzw.apigate.message.error.RestApiException;
 import com.llzw.apigate.persistence.entity.Payment;
 import com.llzw.apigate.persistence.entity.User;
 import com.llzw.apigate.service.PaymentService;
 import com.llzw.apigate.web.dto.PaymentCreateDto;
-import com.llzw.apigate.web.util.RestResponseFactory;
 import java.util.Map;
 import javax.validation.Valid;
 import lombok.Setter;
